refactor(pdf): make PdfOutlineItem an internal type in the outline module
Per review: PdfOutlineItem is an internal backend->heading-stage data structure (already excluded from serialization), not public datamodel. Move it out of base_models into docling/utils/pdf_outline.py, next to the extractors that produce it. Kept as our own type rather than reusing docling-parse's PdfTocEntry, which has no vertical-position (y_top) field and would couple the pypdfium2 backend to docling-parse.
